create_index.py no longer relies on importer/util and thus importer/settings
