+++ labonneboite/importer/conf/development.py
@@ -11,5 +11,3 @@ MAXIMUM_COMPUTE_SCORE_JOB_FAILURES = 94
 SCORE_REDUCING_MINIMUM_THRESHOLD = 0
 
 DEPARTEMENTS_TO_BE_SANITY_CHECKED = ['14', '69']
-
-ENABLE_TIMEIT_TIMERS = False
