+++ labonneboite/importer/conf/test.py
@@ -8,5 +8,3 @@ SCORE_COEFFICIENT_OF_VARIATION_MAX = 1.0
 HIGH_SCORE_COMPANIES_COUNT_MIN = 0
 
 DEPARTEMENTS_TO_BE_SANITY_CHECKED = []
-
-ENABLE_TIMEIT_TIMERS = False
